Update README.md
In this assignment, we will be putting #1 on your resume by completing a STL-compliant HashMap.

Recall that the Map abstract data type stores key value pairs, and provides effcient lookup of the value of any key. There are two versions of this in the STL:→std::map, traditionally implemented using a balanced binary search tree, and std::unordered_map (introduced in C++11), which uses a clever technique called hashing and supports the most map operations in constant time (!).

In this assignment, you will be given a nearly complete implementation of a HashMap (the name of this data structure in the Stanford C++ library). The starter code is nearly complete, but is not “STL-compliant”; for example, copy and move semantics are not supported yet. Your goal is to extend it so it becomes an STL-compliant, industrial strength, robust, and blazingly fast data structure.

A hashmap is a bit of a complicated data structure, but we’ve given you nearly all of the implementation of it.